Update README.md
Financial differences between communities have long been the centre of research and studies, and the dependence of income levels on social factors has led to further disparities between people. This paper tries to examine the likelihood of having high income ($100K or more) on grounds of a person’s gender and level of education. We find that men are significantly more likely to fall in the high income category as compared to women, and that there is a positive correlation between having a higher level of education and earning more. As people go from less than a high school diploma, to some university diploma, to degree above bachelors, we see that their prospects of earning $100K or more also go up.